Moved the filterSortingSidebar to a new file
